test(main): give the Team API suite a harness
Twenty-one cases opened with the same six lines - a temporary directory,
a `TeamStore` over it, initialize, configure an owner, fill in the three
service doubles, start, log in - and then closed with a `try`/`finally`
whose only job was `api.stop()`. `createTeamApiFixture` is those lines,
and `stopTeamApiFixtures` is the `finally`, so a case is now its request
and its assertion.

Building the server is a second call rather than an argument, because
most of what its constructor takes - a `SidebarLayoutStore`, a database,
an agent double closing over a workspace - has to be built inside the
directory first.

Four things the harness will not do on your behalf, each of which would
quietly rewrite what a case tests: default `appVersion` (the protocol
gate returns `null` without one, and the route round-trip sends protocol
headers on none of its ~75 requests, so a default turns that case into
75 x 426), default `logger` (one case silences it deliberately, the
error-path cases need the real call to assert on), start or configure
anything (one case is about `stop`, and identity and join need a store
nobody has configured), or take `now` away from the rate-limit case.

It registers no hook. A module-level `afterEach` here would be
registered before the importing file's, and vitest runs those LIFO, so
the directory would be removed while a listener was still coming down -
the trap `remote-server-test-harness.ts:16-19` already describes. Each
file declares its own `afterEach(stopTeamApiFixtures)`.

Watched it fail: pointing the fixture's `signIn` at a wrong password
reds 15 of the 39 cases with "The username or password is incorrect.",
so the sign-in every one of them now shares is doing the work the
literal it replaced used to do. 39 passed before and after.
